Fix wind model interpolation and test

+++ predict.py
@@ -6,32 +6,57 @@ import ftplib
 import os
 from collections import defaultdict
 import numpy as np
+import math
+
+
+# Constant for Earths radius in miles
+earth_radius = 3960.0
+
+def change_in_latitude(miles):
+    """Given a distance north, return the change in latitude."""
+    return math.degrees(miles / earth_radius)
+
+
+def change_in_longitude(latitude, miles):
+    """Given a latitude and a distance west, return the change in longitude."""
+    # Find the radius of a circle around the earth at given latitude.
+    r = earth_radius * math.cos(math.radians(latitude))
+    return math.degrees(miles / r)
 
 
 class WindModel(object):
-    def __init__(self, center=None, radius=None, forecast_date=None):
+    def __init__(self, center=None, radius=None, forecast_date=None, gfs_data_file=None, hrrr_data_file=None, keep_files=False):
         """
-        Construct a wind model centered at a given Latitude/Longitude with a radius defined in degrees
+        Construct a wind model centered at a given Latitude/Longitude with a radius defined in miles
         :param center:
         :param radius:
         :param forecast_date:
         """
-        if center is None or radius is None or forecast_date is None:
+        if gfs_data_file is None and hrrr_data_file is None and (center is None or radius is None or forecast_date is None):
             raise Exception("Invalid center or radius for wind model")
 
-        self.center = center
-        self.radius = radius
         self.forecast_date = forecast_date
+        self.gfs_data_file = gfs_data_file
         self.gfs_height_map = None
         self.gfs_latlons = None
+        self.hrrr_data_file = hrrr_data_file
         self.hrrr_height_map = None
         self.hrrr_latlons = None
 
-        self.NW_bound = [self.center[0] - self.radius, self.center[1] - self.radius]
-        self.SE_bound = [self.center[0] + self.radius, self.center[1] + self.radius]
+        if gfs_data_file is not None:
+            self.gfs_height_map, self.gfs_latlons = self._parse_grbs(pygrib.open(gfs_data_file))
+        if hrrr_data_file is not None:
+            self.hrrr_height_map, self.hrrr_latlons = self._parse_grbs(pygrib.open(hrrr_data_file))
 
-        self._load_gfs()
-        self._load_hrrr()
+        if gfs_data_file is None and hrrr_data_file is None:
+            lat_radius = change_in_latitude(radius)
+            lon_radius = max(change_in_longitude(center[0] - lat_radius, radius), change_in_longitude(center[0] + lat_radius, radius))
+
+            self.NW_bound = [center[0] + lat_radius, center[1] - lon_radius]
+            self.SE_bound = [center[0] - lat_radius, center[1] + lon_radius]
+
+            self._load_gfs(keep_files)
+            self._load_hrrr(keep_files)
 
     @staticmethod
     def _load_file(url, local_name):
@@ -41,29 +66,43 @@ class WindModel(object):
                 f.write(i)
         return local_name
 
+    NAME_MAP = {
+        'U component of wind' : 'UGRD',
+        'V component of wind': 'VGRD',
+        'Geopotential Height': 'HGT',
+    }
+
     @staticmethod
     def _parse_grbs(grbs):
-        height_map = defaultdict()
-        layer_map = defaultdict(defaultdict())
+        height_map = defaultdict(dict)
         latlons = None
         for grb in grbs:
-            if grb['typeOfLevel'] == 'isobaricInhPa':
-                height_map[grb["layer"]][grb["name"]] = grb.value
+            if grb['typeOfLevel'] == 'isobaricInhPa' and 'level' in grb.keys():
+                data_name = WindModel.NAME_MAP[grb['name']]
+                height_map[grb['level']][data_name] = grb.values
+                if data_name == 'HGT':
+                    height_map[grb['level']]['HGT-AVG'] = grb['avg']
                 if latlons is None: latlons = grb.latlons()
-        for level in layer_map:
-            layer = layer_map[level]
-            height_map[layer["HGT"]] = {"UGRD":layer["UGRD"], "VGRD":layer["VGRD"], "hPa":layer}
+        latlons = (latlons[0],latlons[1]-360)
         return height_map, latlons
 
-    def _load_generic(self, ftp_dir, prefix, level_type, expected_forecast, filter_type):
+    def _load_generic(self, ftp_dir, prefix, level_type, expected_forecast, filter_type, local_file, keep_file=False):
         try:
             ftp = ftplib.FTP("ftp.ncep.noaa.gov")
             ftp.login()
             ftp.cwd(ftp_dir)
-            recent = max((int(f.replace(prefix, "")), f) for f in ftp.nlst() if prefix in f)[1]
+            run_days = [(int(f.replace(prefix, "")), f) for f in ftp.nlst() if prefix in f]
+            recent = max(run_days)[1]
             ftp.cwd(recent)
             forecasts = [f for f in ftp.nlst() if level_type in f and not f.endswith(".idx")]
-            selected_forecast = [x for x in forecasts if 'f' + str(expected_forecast).zfill(3) in x][0]
+            forecast_postfix = 'f' + str(int(expected_forecast)).zfill(3)
+            if not any(forecast_postfix in x for x in forecasts):
+                ftp.cwd("..")
+                run_days.remove(max(run_days))
+                recent = max(run_days[1])
+                ftp.cwd(recent)
+                forecasts = [f for f in ftp.nlst() if level_type in f and not f.endswith(".idx")]
+            selected_forecast = [x for x in forecasts if forecast_postfix in x][0]
             params = [
                 "file=%s",
                 "all_lev=on",
@@ -78,40 +117,46 @@ class WindModel(object):
                 "dir=%s"
             ]
             constructed_url = "http://nomads.ncep.noaa.gov/cgi-bin/%s?" + "&".join(params)
-            constructed_url = constructed_url % (filter_type, selected_forecast, self.NW_bound[1], self.SE_bound[1], self.NW_bound[0], self.SE_bound[0], recent)
-            print constructed_url
-            local_file = self._load_file(constructed_url, prefix + str(time.time()) + ".grib2")
+            constructed_url = constructed_url % (
+                filter_type, selected_forecast, self.NW_bound[1], self.SE_bound[1], self.NW_bound[0], self.SE_bound[0],
+                "%2F"+recent)
+            self._load_file(constructed_url, local_file)
             grbs = pygrib.open(local_file)
             return self._parse_grbs(grbs)
         except:
             raise
         finally:
             try:
-                os.remove(local_file)
-            except: pass
+                if not keep_file:
+                    os.remove(local_file)
+            except:
+                pass
 
-
-    def _load_gfs(self):
-        expected_forecast = max(min((self.forecast_date - datetime.datetime.utcnow()).total_seconds() / 3600, 384),0)
+    def _load_gfs(self, keep_files):
+        expected_forecast = max(min(int((self.forecast_date - datetime.datetime.utcnow()).total_seconds() / 3600), 384), 0)
         if expected_forecast >= 120: expected_forecast = expected_forecast / 3 * 3
-        self.gfs_height_map, self.gfs_latlons = self._load_generic("pub/data/nccf/com/gfs/prod", "gfs.", "pgrb2.0p25", expected_forecast,"filter_gfs_0p25.pl")
+        local_file = os.path.join("saved","gfs." + str(time.time()) + ".grib2")
+        self.gfs_height_map, self.gfs_latlons = self._load_generic("pub/data/nccf/com/gfs/prod", "gfs.", "pgrb2.0p25",
+                                                                   expected_forecast, "filter_gfs_0p25.pl", local_file, keep_file=keep_files)
+        if keep_files:
+            self.gfs_data_file = local_file
 
-    def _load_hrrr(self):
-        expected_forecast = max((self.forecast_date - datetime.datetime.utcnow()).total_seconds() / 3600,0)
+    def _load_hrrr(self, keep_files):
+        expected_forecast = max(int((self.forecast_date - datetime.datetime.utcnow()).total_seconds() / 3600), 0)
         if expected_forecast > 18: return
-        self.hrrr_height_map, self.hrrr_latlons = self._load_generic("pub/data/nccf/com/hrrr/prod", "hrrr.", "wrfprs", expected_forecast,"filter_hrrr_2d.pl")
+        local_file = os.path.join("saved","hrrr." + str(time.time()) + ".grib2")
+        self.hrrr_height_map, self.hrrr_latlons = self._load_generic("pub/data/nccf/com/hrrr/prod", "hrrr.", "wrfprs",
+                                                                     expected_forecast, "filter_hrrr_2d.pl", local_file, keep_file=keep_files)
+        if keep_files:
+            self.hrrr_data_file = local_file
 
     @staticmethod
     def get_closest_bounds(val, iterable):
         if iterable is None:
             return 0, 0
-        below = 0
-        above = 0
-        for level in iterable:
-            if level < val:
-                below = level
-            if above <= below:
-                above = level
+        arr = np.array(iterable)
+        below = arr[arr <= val].max()
+        above = arr[arr > val].min()
         return below, above
 
     @staticmethod
@@ -119,7 +164,9 @@ class WindModel(object):
         if height_map is None or latlons is None:
             return None, None
 
-        level_below, level_above = WindModel.get_closest_bounds(alt, height_map)
+        alt_levels = WindModel.get_closest_bounds(alt, [height_map[x]["HGT-AVG"] for x in height_map])
+
+        level_below, level_above = [level for x in alt_levels for level in height_map if height_map[level]["HGT-AVG"] == x]
 
         if level_below == level_above:
             return None, None
@@ -128,18 +175,19 @@ class WindModel(object):
             return float(y_min * (x_max - x) + y_max * (x - x_min)) / float(x_max - x_min)
 
         def interp_level(layer):
-            lats = latlons[0][:,0]
-            lons = latlons[1][0,:]
+            lats = latlons[0][:, 0]
+            lons = latlons[1][0, :]
 
             lat_bounds = WindModel.get_closest_bounds(lat, lats)
-            lat_idx = [np.where(lats==lat_bounds[i])[0][0] for i in range(2)]
+            lat_idx = [np.where(lats == lat_bounds[i])[0][0] for i in range(2)]
             lon_bounds = WindModel.get_closest_bounds(lon, lons)
             lon_idx = [np.where(lons == lon_bounds[i])[0][0] for i in range(2)]
 
             def lat_lerp(key, tmp_lon_idx):
-                return lerp(lat, lat_bounds[0], lat_bounds[1], layer[key][lat_idx[0],tmp_lon_idx],layer[key][lat_idx[1],tmp_lon_idx])
+                return lerp(lat, lat_bounds[0], lat_bounds[1], layer[key][lat_idx[0], tmp_lon_idx],
+                            layer[key][lat_idx[1], tmp_lon_idx])
 
-            lat_interp_0 = [lat_lerp("UGRD",lon_idx[0]),lat_lerp("VGRD",lon_idx[0])]
+            lat_interp_0 = [lat_lerp("UGRD", lon_idx[0]), lat_lerp("VGRD", lon_idx[0])]
             lat_interp_1 = [lat_lerp("UGRD", lon_idx[1]), lat_lerp("VGRD", lon_idx[1])]
 
             return [lerp(lon, lon_bounds[0], lon_bounds[1], lat_interp_0[i], lat_interp_1[i]) for i in range(2)]
@@ -147,7 +195,7 @@ class WindModel(object):
         interp_above = interp_level(height_map[level_above])
         interp_below = interp_level(height_map[level_below])
 
-        return [lerp(alt, level_below, level_above, interp_below[i], interp_above[i]) for i in range(2)]
+        return [lerp(alt, alt_levels[0], alt_levels[1], interp_below[i], interp_above[i]) for i in range(2)]
 
     def get_winds(self, lat, lon, alt):
         """
@@ -159,7 +207,8 @@ class WindModel(object):
         """
         alt_m = alt * 0.3048
 
-        ugrd, vgrd = self.get_wind_from_map(lat, lon, alt_m, self.hrrr_height_map, self.hrrr_latlons)
+        ugrd, vgrd = None, None
+        # ugrd, vgrd = self.get_wind_from_map(lat, lon, alt_m, self.hrrr_height_map, self.hrrr_latlons)
         if ugrd is None or vgrd is None:
             ugrd, vgrd = self.get_wind_from_map(lat, lon, alt_m, self.gfs_height_map, self.gfs_latlons)
 
@@ -167,3 +216,13 @@ class WindModel(object):
             raise Exception("Received invalid gfs levels")
 
         return [ugrd * 3.28084, vgrd * 3.28084]
+
+
+if __name__ == "__main__":
+    # Test
+    center = [43.98824490544571, -112.73088455200195]
+    radius = 100 # miles
+    forecast_date = datetime.datetime.utcnow() + datetime.timedelta(6)
+    # model = WindModel(center=center, radius=radius, forecast_date=forecast_date, keep_files=True)
+    model = WindModel(gfs_data_file="saved/gfs.1498452054.93.grib2")
+    print model.get_winds(43.816442,-111.7459983, 5000)
